apollo_fpga.gateware: add ApolloAdvertiser
When this is added as a submodule to a design, an advertisement message
is sent periodically to Apollo. Apollo will take over the port when
these announcements are interrupted or when the PROGRAM button is
pressed.

An optional request handler is added: REQUEST_APOLLO_ADV_STOP (0xf0).
It returns the CONTROL port to Apollo by stopping announcements.