apollo_fpga: Ignore devices not running Apollo
When looking for devices running Apollo firmware, ignore those running
Saturn-V or anything else that does not identify itself as "Apollo" with
a vendor request. Because the USB VID/PID is shared between Saturn-V and
Apollo, this is needed to prevent us from trying to communicate with a
Saturn-V device as if it were running Apollo.